+++ css-base/src/main/java/org/spc/base/client/ProcessClient.java
@@ -28,4 +28,10 @@ public interface ProcessClient {
      */
     @PostMapping("/process/commandExecution/{order}")
     void commandExecution(@PathVariable String order);
+
+    /**
+     * 运行进程
+     */
+    @PostMapping("/process/runProcess/")
+    void runProcess(Process process);
 }
