+++ css-file/src/main/java/org/spc/file/api/FileApi.java
@@ -1,0 +1,19 @@
+package org.spc.file.api;
+
+
+import lombok.RequiredArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
+
+/**
+ * 文件Controller
+ */
+@Slf4j
+@RestController
+@RequestMapping("/file")
+@RequiredArgsConstructor
+public class FileApi {
+
+
+}
